requirements.txt
aiogram==3.7.0
playwright==1.44.0
vk-api==11.9.9
apscheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.5
